MinorUpdt

+++ internal/infra/http/controllers/house_controller.go
@@ -55,7 +55,13 @@ func (c HouseController) Update() http.HandlerFunc {
 			return
 		}
 
+		user := r.Context().Value(UserKey).(domain.User)
 		house := r.Context().Value(HouseKey).(domain.House)
+		if user.Id != house.UserId {
+			err := errors.New("Acces denied")
+			Forbidden(w, err)
+			return
+		}
 
 		house, err = c.houseService.Update(updt, house)
 		if err != nil {
@@ -73,6 +79,12 @@ func (c HouseController) Update() http.HandlerFunc {
 func (c HouseController) Delete() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		house := r.Context().Value(HouseKey).(domain.House)
+		user := r.Context().Value(UserKey).(domain.User)
+		if house.UserId != user.Id {
+			err := errors.New("Acces denied")
+			Forbidden(w, err)
+			return
+		}
 
 		err := c.houseService.Delete(house.Id)
 		if err != nil {
@@ -95,6 +107,14 @@ func (c HouseController) Find() http.HandlerFunc {
 			Forbidden(w, err)
 			return
 		}
+
+		rooms, err := c.houseService.FindbyIdRooms(house.Id)
+		if err != nil {
+			log.Printf("Error,c HouseController.Find().FindbyIdRooms(): %s", err)
+			return
+		}
+
+		house.Rooms = rooms
 
 		var houseDto resources.HouseDto
 		houseDto = houseDto.DomainToDto(house)
